+++ CHANGELOG.md
@@ -42,6 +42,7 @@ described here [Making a flat list out of lists of lists](http://stackoverflow.c
 ## Version 0.2.3 (2015-02-13)
 
 ### Features
+* Added `skbio.stats.composition` for analyzing data made up of proportions
 * Modified ``skbio.stats.distance.pwmantel`` to accept a list of filepaths. This is useful as it allows for a smaller amount of memory consumption as it only loads two matrices at a time as opposed to requiring that all distance matrices are loaded into memory.
 * Added ``skbio.util.find_duplicates`` for finding duplicate elements in an iterable.